config: move the inline test modules into their own files
Pure move, zero behaviour change. The config module was the outlier: the
tree already has ten separate test directories, `config/tests/` among
them, and `migrate.rs`, `mod.rs` and `named_map.rs` already used the
`#[cfg(test)] #[path = "tests/..."] mod ...;` declaration form. Eight
inline test suites did not: groups.rs, patch.rs, overlay.rs (three) and
secret.rs (three).

WHY IT MATTERS, and it is not taste. The rationale is that it keeps file
length honest, which is a property of every judgement anyone makes about
this tree from a line count. It had already corrupted one: overlay.rs was
reported at 2,111 lines and reasoned about against migrate.rs at 2,379,
concluding things about which file was more bloated. The real overlay
figure was 607 lines of code and 919 of tests. The comparison that
started that whole thread was an artifact of the violation. It is not
academic either: structure-lint.sh's 2,500-line ceiling and its
grandfathered-oversized ledger are computed from exactly that
measurement.

overlay.rs is now 1,086 lines and secret.rs 345.

PROVEN A PURE MOVE, both ways, because a move that silently drops a test
is worse than not moving it and a smaller number in a wall of green is
easy to miss:
  * the full list of test NAMES before and after is byte-identical,
    3,406 either side. Nothing added, dropped or renamed.
  * the full suite is green: 3,078 in the engine plus every sibling
    target, 0 failures.

No visibility was weakened to do it. Every moved test compiled against
`use super::*` from a `#[path]` child module, which reaches exactly what
the inline module reached, so no item needed widening to `pub(crate)` and
no test needed an accessor.

THE NAME DIFF ALONE WAS NOT ENOUGH, which is worth recording. A first
attempt dedented each moved body by four columns to undo the module
indentation. That silently rewrote the inside of a multi-line YAML string
LITERAL in `group_yaml_round_trips_exactly` -- the `child_default:` block
lost its indentation and stopped parsing. The test-name diff was clean;
only running the suite caught it. The move now copies bodies VERBATIM and
lets `cargo fmt` reindent, which is the one tool that knows the
difference between code and the inside of a string.
