+++ crates/busbar/src/config/secret.rs
@@ -333,340 +333,13 @@ pub(crate) fn resolve_builtin_string(secret: &SecretRef) -> Result<String, Strin
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    /// The env built-in resolves a set variable, trims trailing newlines in the string form, and
-    /// fails closed on unset / empty values.
-    #[test]
-    fn env_module_resolves_and_fails_closed() {
-        let var = format!("BUSBAR_SECRET_TEST_{}", std::process::id());
-        std::env::set_var(&var, "s3cret\n");
-        let r = SecretRef::env(&var);
-        assert_eq!(resolve_builtin(&r).unwrap(), b"s3cret\n");
-        assert_eq!(resolve_builtin_string(&r).unwrap(), "s3cret");
-        std::env::remove_var(&var);
-        let err = resolve_builtin(&r).unwrap_err();
-        assert!(err.contains("unset"), "unset env is fail-closed: {err}");
-        std::env::set_var(&var, "");
-        let err = resolve_builtin(&r).unwrap_err();
-        assert!(err.contains("EMPTY"), "empty env is fail-closed: {err}");
-        std::env::remove_var(&var);
-    }
-
-    /// The file built-in resolves file bytes and fails closed on a missing or empty file.
-    #[test]
-    fn file_module_resolves_and_fails_closed() {
-        let path = std::env::temp_dir().join(format!("busbar-secret-{}.txt", std::process::id()));
-        std::fs::write(&path, b"file-secret\n").unwrap();
-        let r = SecretRef::file(path.to_string_lossy().into_owned());
-        assert_eq!(resolve_builtin(&r).unwrap(), b"file-secret\n");
-        assert_eq!(resolve_builtin_string(&r).unwrap(), "file-secret");
-        std::fs::write(&path, b"").unwrap();
-        let err = resolve_builtin(&r).unwrap_err();
-        assert!(err.contains("EMPTY"), "empty file is fail-closed: {err}");
-        let _ = std::fs::remove_file(&path);
-        let err = resolve_builtin(&r).unwrap_err();
-        assert!(err.contains("cannot resolve"), "missing file fails: {err}");
-    }
-
-    /// A whitespace-only (or empty) `env:` NAME — not value — must be rejected at the settings-shape
-    /// layer (`self_env_var_checked`), never silently treated as "no name given, fall through" or
-    /// "look up an env var literally named '   '". Distinct from `env_module_resolves_and_fails_closed`
-    /// above, which covers the RESOLVED VALUE being empty, not the configured variable name itself.
-    #[test]
-    fn env_whitespace_only_name_is_rejected() {
-        for bad in ["", "   ", "\t\n"] {
-            let r = SecretRef::env(bad);
-            let err = resolve_builtin(&r).expect_err(&format!(
-                "whitespace-only env name {bad:?} must be rejected"
-            ));
-            assert!(
-                err.contains("requires settings.key"),
-                "whitespace-only env name {bad:?} must fail the settings-shape check, got: {err}"
-            );
-        }
-    }
-
-    /// Same guard, `file:` PATH side.
-    #[test]
-    fn file_whitespace_only_path_is_rejected() {
-        for bad in ["", "   ", "\t\n"] {
-            let r = SecretRef::file(bad);
-            let err = resolve_builtin(&r).expect_err(&format!(
-                "whitespace-only file path {bad:?} must be rejected"
-            ));
-            assert!(
-                err.contains("requires settings.path"),
-                "whitespace-only file path {bad:?} must fail the settings-shape check, got: {err}"
-            );
-        }
-    }
-
-    /// An unknown secret module is FAIL-CLOSED at the built-in resolver (the plugin-backed
-    /// resolver layers on top; anything it cannot resolve lands here and refuses).
-    #[test]
-    fn unknown_module_fails_closed() {
-        let r = SecretRef {
-            module: "vault".to_string(),
-            settings: serde_json::Map::new(),
-        };
-        let err = resolve_builtin(&r).unwrap_err();
-        assert!(
-            err.contains("fail-closed") && err.contains("vault"),
-            "unknown module refuses: {err}"
-        );
-    }
-
-    /// Malformed built-in refs (env without key, file without path) error precisely, never fall
-    /// through to "unknown module".
-    #[test]
-    fn malformed_builtin_refs_error_precisely() {
-        let r = SecretRef {
-            module: SECRET_MODULE_ENV.to_string(),
-            settings: serde_json::Map::new(),
-        };
-        assert!(resolve_builtin(&r).unwrap_err().contains("settings.key"));
-        let r = SecretRef {
-            module: SECRET_MODULE_FILE.to_string(),
-            settings: serde_json::Map::new(),
-        };
-        assert!(resolve_builtin(&r).unwrap_err().contains("settings.path"));
-    }
-
-    /// Deserialize: the `{env}` / `{file}` sugar desugars to the canonical module + settings; the
-    /// canonical form parses; mixed / unknown / empty forms are rejected.
-    #[test]
-    fn deserialize_accepts_canonical_and_sugar_rejects_malformed() {
-        let r: SecretRef = serde_yaml::from_str("{ env: MY_VAR }").unwrap();
-        assert_eq!(r, SecretRef::env("MY_VAR"));
-        assert_eq!(r.env_var(), Some("MY_VAR"));
-        let r: SecretRef = serde_yaml::from_str("{ file: /run/secrets/x }").unwrap();
-        assert_eq!(r, SecretRef::file("/run/secrets/x"));
-        assert_eq!(r.file_path(), Some("/run/secrets/x"));
-        let r: SecretRef =
-            serde_yaml::from_str("{ module: vault, settings: { path: kv/data/x } }").unwrap();
-        assert_eq!(r.module, "vault");
-        assert_eq!(
-            r.settings.get("path").and_then(|v| v.as_str()),
-            Some("kv/data/x")
-        );
-
-        for bad in [
-            "{ env: A, file: B }",
-            "{ module: vault, env: A }",
-            "{ env: A, settings: {} }",
-            "{ unknown_key: A }",
-            "{}",
-            "{ env: \"\" }",
-            "{ module: \"\" }",
-            "plain-string",
-        ] {
-            assert!(
-                serde_yaml::from_str::<SecretRef>(bad).is_err(),
-                "must reject: {bad}"
-            );
-        }
-    }
-}
+#[path = "tests/secret_tests.rs"]
+mod tests;
 
 #[cfg(test)]
-mod resolver_tests {
-    use super::*;
-
-    /// The resolver's built-in arm resolves env/file inline WITHOUT any plugin.
-    #[test]
-    fn resolver_builtins_resolve_without_plugin() {
-        let r = SecretResolver::builtins_only();
-        let var = format!("BUSBAR_RESOLVER_TEST_{}", std::process::id());
-        std::env::set_var(&var, "abc\n");
-        assert_eq!(r.resolve(&SecretRef::env(&var)).unwrap(), b"abc\n");
-        assert_eq!(r.resolve_string(&SecretRef::env(&var)).unwrap(), "abc");
-        std::env::remove_var(&var);
-    }
-
-    /// A non-built-in module with NO plugin subsystem is FAIL-CLOSED, naming the module.
-    #[test]
-    fn resolver_unknown_module_without_plugin_fails_closed() {
-        let r = SecretResolver::builtins_only();
-        let s = SecretRef {
-            module: "vault".to_string(),
-            settings: serde_json::Map::new(),
-        };
-        let err = r.resolve(&s).unwrap_err();
-        assert!(
-            err.contains("vault") && err.contains("fail-closed"),
-            "unknown module with no plugin refuses: {err}"
-        );
-    }
-
-    /// A non-built-in module DELEGATES to the plugin resolver; a plugin error and an empty result
-    /// are both fail-closed.
-    #[test]
-    fn resolver_delegates_to_plugin_and_fails_closed_on_empty_or_error() {
-        // Plugin that returns bytes for `vault` and errors for anything else.
-        let r = SecretResolver::with_plugin(Box::new(|module: &str, settings: &str| {
-            if module == "vault" {
-                let v: serde_json::Value = serde_json::from_str(settings).unwrap();
-                match v.get("path").and_then(|p| p.as_str()) {
-                    Some("kv/ok") => Ok(b"plugin-secret".to_vec()),
-                    Some("kv/empty") => Ok(Vec::new()),
-                    _ => Err("no such path".to_string()),
-                }
-            } else {
-                Err("unknown module".to_string())
-            }
-        }));
-        let mk = |path: &str| {
-            let mut settings = serde_json::Map::new();
-            settings.insert(
-                "path".to_string(),
-                serde_json::Value::String(path.to_string()),
-            );
-            SecretRef {
-                module: "vault".to_string(),
-                settings,
-            }
-        };
-        assert_eq!(r.resolve(&mk("kv/ok")).unwrap(), b"plugin-secret");
-        // Empty plugin result is rejected (fail-closed), never an empty secret.
-        assert!(r.resolve(&mk("kv/empty")).unwrap_err().contains("EMPTY"));
-        // Plugin error is surfaced fail-closed.
-        assert!(r
-            .resolve(&mk("kv/missing"))
-            .unwrap_err()
-            .contains("failed to resolve"));
-        // Built-ins still short-circuit past the plugin.
-        let var = format!("BUSBAR_RESOLVER_PLUGIN_TEST_{}", std::process::id());
-        std::env::set_var(&var, "envval");
-        assert_eq!(r.resolve_string(&SecretRef::env(&var)).unwrap(), "envval");
-        std::env::remove_var(&var);
-    }
-}
+#[path = "tests/secret_resolver_tests.rs"]
+mod resolver_tests;
 
 #[cfg(test)]
-mod settings_resolution_tests {
-    use super::*;
-
-    fn obj(pairs: &[(&str, serde_json::Value)]) -> serde_json::Map<String, serde_json::Value> {
-        pairs
-            .iter()
-            .map(|(k, v)| (k.to_string(), v.clone()))
-            .collect()
-    }
-
-    /// A SecretRef-shaped setting (the `{ env: … }` sugar for a `licenseKey`) is RESOLVED to its
-    /// value before delivery, while a plain sibling setting passes through verbatim.
-    #[test]
-    fn resolves_secret_ref_settings_and_passes_plain_through() {
-        let var = format!("BUSBAR_PLUGIN_LICENSE_TEST_{}", std::process::id());
-        std::env::set_var(&var, "LIC-123\n");
-        let resolver = SecretResolver::builtins_only();
-        let settings = obj(&[
-            ("licenseKey", serde_json::json!({ "env": var })),
-            ("endpoint", serde_json::json!("https://example.test")),
-            ("retries", serde_json::json!(3)),
-        ]);
-        let out = resolve_settings(&settings, &resolver).unwrap();
-        // The ref resolved to the raw secret (trailing newline trimmed) — a plain STRING now.
-        assert_eq!(
-            out.get("licenseKey").unwrap(),
-            &serde_json::json!("LIC-123")
-        );
-        // Non-ref settings are untouched.
-        assert_eq!(
-            out.get("endpoint").unwrap(),
-            &serde_json::json!("https://example.test")
-        );
-        assert_eq!(out.get("retries").unwrap(), &serde_json::json!(3));
-        std::env::remove_var(&var);
-    }
-
-    /// An ordinary settings OBJECT that is not a secret reference (its keys aren't a ref's keys) is
-    /// left untouched — resolution only fires on a genuine ref shape.
-    #[test]
-    fn leaves_non_ref_object_settings_untouched() {
-        let resolver = SecretResolver::builtins_only();
-        let settings = obj(&[("db", serde_json::json!({ "path": ":memory:", "wal": true }))]);
-        let out = resolve_settings(&settings, &resolver).unwrap();
-        assert_eq!(
-            out.get("db").unwrap(),
-            &serde_json::json!({ "path": ":memory:", "wal": true })
-        );
-    }
-
-    /// FAIL-CLOSED: a SecretRef setting that cannot resolve (unset env) is a hard error naming the
-    /// FIELD — never a silently-empty or dangling value handed to the plugin. The error text does not
-    /// echo any secret value.
-    #[test]
-    fn unresolvable_secret_ref_setting_fails_closed_naming_field() {
-        let var = format!("BUSBAR_PLUGIN_LICENSE_MISSING_{}", std::process::id());
-        std::env::remove_var(&var);
-        let resolver = SecretResolver::builtins_only();
-        let settings = obj(&[("license", serde_json::json!({ "env": var }))]);
-        let err = resolve_settings(&settings, &resolver).unwrap_err();
-        assert!(err.contains("license"), "names the field: {err}");
-        assert!(err.contains("did not resolve"), "fail-closed: {err}");
-    }
-
-    /// An unknown secret module in a plugin setting is FAIL-CLOSED when no secret plugin can resolve
-    /// it (built-ins-only resolver), never handed through.
-    #[test]
-    fn unknown_secret_module_in_setting_fails_closed() {
-        let resolver = SecretResolver::builtins_only();
-        let settings = obj(&[(
-            "licenseKey",
-            // settings-leak-lint: allow — test fixture: a `kind: secret` module reference, not a response body.
-            serde_json::json!({ "module": "vault", "settings": { "path": "kv/license" } }),
-        )]);
-        let err = resolve_settings(&settings, &resolver).unwrap_err();
-        assert!(
-            err.contains("licenseKey") && err.contains("fail-closed"),
-            "unknown module fails closed: {err}"
-        );
-    }
-
-    /// #40: a plugin setting shaped like a reference is AMBIGUOUS by nature, and the coercion used
-    /// to be silent — a plugin whose own config is `{ file: /var/lib/db }` (a PATH it opens) or
-    /// `{ env: HOME }` (a variable NAME it reads) had the value swapped for the file's contents /
-    /// the variable's value, with no diagnostic. `{ literal: … }` is the escape hatch: the inner
-    /// value is delivered verbatim and never resolved.
-    #[test]
-    fn literal_wrapper_opts_a_setting_out_of_secret_coercion() {
-        let path = std::env::temp_dir().join(format!("busbar-lit-{}.txt", std::process::id()));
-        std::fs::write(&path, b"THE-SECRET").unwrap();
-        let resolver = SecretResolver::builtins_only();
-
-        // Un-wrapped, the ref shape IS coerced (the documented ADR-0010 behaviour).
-        let mut settings = serde_json::Map::new();
-        settings.insert(
-            "licenseKey".to_string(),
-            serde_json::json!({ "file": path.to_string_lossy() }),
-        );
-        let out = resolve_settings(&settings, &resolver).expect("resolves");
-        assert_eq!(out["licenseKey"], serde_json::json!("THE-SECRET"));
-
-        // WRAPPED, the very same object is delivered verbatim — the plugin sees its own config.
-        let mut settings = serde_json::Map::new();
-        settings.insert(
-            "db".to_string(),
-            serde_json::json!({ "literal": { "file": path.to_string_lossy() } }),
-        );
-        let out = resolve_settings(&settings, &resolver).expect("resolves");
-        assert_eq!(
-            out["db"],
-            serde_json::json!({ "file": path.to_string_lossy() }),
-            "a `literal:`-wrapped object is passed through untouched, never dereferenced"
-        );
-
-        // The wrapper is not limited to ref-shaped objects, and it is exactly one level deep.
-        let mut settings = serde_json::Map::new();
-        settings.insert("n".to_string(), serde_json::json!({ "literal": 42 }));
-        settings.insert("plain".to_string(), serde_json::json!({ "db_path": "x" }));
-        let out = resolve_settings(&settings, &resolver).expect("resolves");
-        assert_eq!(out["n"], serde_json::json!(42));
-        assert_eq!(out["plain"], serde_json::json!({ "db_path": "x" }));
-
-        let _ = std::fs::remove_file(&path);
-    }
-}
+#[path = "tests/secret_settings_resolution_tests.rs"]
+mod settings_resolution_tests;
